Crate tsukuyomi

Tsukuyomi is a next generation Web framework for Rust.

Modules

app
Components for constructing HTTP applications.

error
Components for constructing and handling HTTP errors.

handler
Handler and supplemental components.

input
Components for parsing incoming HTTP requests and accessing the global or request-local data.

json
Components for parsing JSON values and creating JSON responses.

local
A testing framework for Tsukuyomi.

modifier
Modifier and supplemental components.

output
Components for constructing HTTP responses.

server
The implementation of low level HTTP server.

Macros

local_key
A macro to create a LocalKey<T>.

Structs

App
The main type which represents an HTTP application.

Error
A type which holds all kinds of errors occurring in handlers.

Input
Contextual information used by processes during an incoming HTTP request.

Traits

AsyncResponder
The async variant of Responder.

Handler
A trait representing handler functions.

Modifier
A trait representing a Modifier.

Responder
A trait representing the conversion to an HTTP response.

Functions

run
Starts an HTTP server with a constructed App and the default server configuration.

Type Definitions

AppResult
A type alias of Result<T, E> which will be returned from run.

Output
The type representing outputs returned from handlers.

Result
A type alias of Result<T, E> with error::Error as error type.