Struct tsukuyomi::app::builder::Scope

pub struct Scope<'a> { /* fields omitted */ }

A proxy object for configuration of a scope.

Methods

impl<'a> Scope<'a>

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn route(
&mut self,
config: impl RouteConfig
) -> &mut Self

Adds a route into the current scope, with the provided configuration.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn scope(
&mut self,
config: impl ScopeConfig
) -> &mut Self

Create a new sub-scope with the provided configuration.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn mount(
&mut self,
prefix: &str,
f: impl FnOnce(&mut Scope)
) -> &mut Self

Create a new scope mounted to the certain URI.

This method is a shortcut of Scope::scope(Mount(prefix, f)).

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn set<T>(&mut self, value: T) -> &mut Self where
T: Send + Sync + 'static,

Adds a scope-local variable into the application.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn prefix(&mut self, prefix: &str) -> &mut Self

Modifies the prefix URI of current scope.

Important traits for &'a mut R
impl<'a, R> Read for &'a mut R where
R: Read + ?Sized, impl<'a, W> Write for &'a mut W where
W: Write + ?Sized, impl<'a, I> Iterator for &'a mut I where
I: Iterator + ?Sized, type Item = <I as Iterator>::Item;

pub fn modifier(
&mut self,
modifier: impl Modifier + Send + Sync + 'static
) -> &mut Self

Register a Modifier into the current scope.