Struct tsukuyomi::app::config::Resource

pub struct Resource<'s, P, M, C> where
    P: Path,
    C: Concurrency,  { /* fields omitted */ }

A resource associated with a specific HTTP path.

Methods

impl<'s, P, M, C> Resource<'s, P, M, C> where
P: Path,
C: Concurrency,

pub fn route(
&mut self,
methods: impl IntoIterator<Item = impl Into<Method>>
) -> Route<PathExtractor<P>, &M, C>

Creates a Route that matches to the specified HTTP methods.

pub fn get(&mut self) -> Route<PathExtractor<P>, &M, C>

pub fn post(&mut self) -> Route<PathExtractor<P>, &M, C>

pub fn put(&mut self) -> Route<PathExtractor<P>, &M, C>

pub fn head(&mut self) -> Route<PathExtractor<P>, &M, C>

pub fn delete(&mut self) -> Route<PathExtractor<P>, &M, C>

pub fn patch(&mut self) -> Route<PathExtractor<P>, &M, C>

pub fn any(&mut self) -> Route<PathExtractor<P>, &M, C>

Start building of a Route that matches any HTTP method.

pub fn to<T>(&mut self, endpoint: T) -> Result<()> where
T: Endpoint<P::Output>,
M: ModifyHandler<EndpointHandler<PathExtractor<P>, T>>,
M::Handler: Into<C::Handler>,

pub fn with<M2>(self, modifier: M2) -> Resource<'s, P, Chain<M2, M>, C>

Appends a ModifyHandler to the stack applied to the all handlers on this resource.

pub fn done<F, T>(self, f: F) -> Result<T> where
F: FnOnce(Self) -> Result<T>,

Applies itself to the specified function.