Module tsukuyomi::extractor

Definition of Extractor and its implementors.

Re-exports

pub use self::ext::ExtractorExt;

Modules

body
Extractors for parsing message body.

ext
A set of extensions for Extractors.

header
Extractors for accessing HTTP header fields.

local
Extractors for accessing the request-local data.

method
Extractors for validation of HTTP method.

Traits

Extractor
A trait abstracting the extraction of values from the incoming request.

Functions

extension
Creates an Extractor that returns the value of extension of the specified type.

extract
Creates an Extractor from the provided function that returns a TryFuture.

method
Creates an Extractor that returns the value of request method.

query
Creates an Extractor that parses the value of query string to T.

ready
Creates an Extractor from a synchronous function.

uri
Creates an Extractor that returns the value of request URI.

value
Creates an Extractor that just clones and returns the provided value.

version
Creates an Extractor that returns the value of the request version.