Trait tsukuyomi::input::localmap::LocalData

pub trait LocalData: Sized + Send + 'static {
    const KEY: LocalKey<Self>;
    fn get(map: &LocalMap) -> Option<&Self> { ... }

    fn get_mut(map: &mut LocalMap) -> Option<&mut Self> { ... }

    fn contains(map: &LocalMap) -> bool { ... }

    fn take_from(map: &mut LocalMap) -> Option<Self> { ... }

    fn insert_into(self, map: &mut LocalMap) -> Option<Self> { ... }

    fn entry(map: &mut LocalMap) -> Entry<Self> { ... }
}

A trait representing a data to be stored in LocalMaps.

Associated Constants

const KEY: LocalKey<Self>

The value of LocalKey associated with this type.

Loading content...
Provided methods

fn get(map: &LocalMap) -> Option<&Self>

Acquires a reference to Self stored in the specified localmap.

fn get_mut(map: &mut LocalMap) -> Option<&mut Self>

Acquires a mutable reference to Self stored in the specified localmap.

fn contains(map: &LocalMap) -> bool

Returns true if a value of Self is stored in the specified localmap.

fn take_from(map: &mut LocalMap) -> Option<Self>

Extracts the instance of Self stored in the specified localmap.

fn insert_into(self, map: &mut LocalMap) -> Option<Self>

Stores itself into the specified localmap.

fn entry(map: &mut LocalMap) -> Entry<Self>

Creates an Entry for in-place manipulation corresponds to an entry of Self in the specified map.

Loading content...