Struct tsukuyomi::input::Input

pub struct Input<'task> { /* fields omitted */ }

Contextual information used by processes during an incoming HTTP request.

Methods

impl<'task> Input<'task>

pub fn request(&self) -> &Request<RequestBody>

Returns a shared reference to the value of Request contained in this context.

pub fn request_mut(&mut self) -> &mut Request<RequestBody>

Returns a mutable reference to the value of Request contained in this context.

pub fn params(&self) -> Params

Returns a proxy object for accessing parameters extracted by the router.

pub fn get<T>(&self) -> Option<&T> where
T: Send + Sync + 'static,

Returns the reference to a value of T registered in the global storage, if possible.

This method will return a None if a value of T is not registered in the global storage.

pub fn cookies(&mut self) -> Result<&mut CookieJar, Error>

Returns a proxy object for managing the value of Cookie entries.

This function will perform parsing when called at first, and returns an Err if the value of header field is invalid.

pub fn locals(&self) -> &LocalMap

Returns a reference to LocalMap for managing request-local data.

pub fn locals_mut(&mut self) -> &mut LocalMap

Returns a mutable reference to LocalMap for managing request-local data.