Trait tsukuyomi::json::HttpResponse

pub trait HttpResponse {
    fn status_code(&self) -> StatusCode { ... }

    fn append_headers(&self, headers: &mut HeaderMap) { ... }
}

A trait representing additional information for constructing HTTP responses from Json<T>.

Provided Methods

fn status_code(&self) -> StatusCode

Returns an HTTP status code associated with the value of this type.

fn append_headers(&self, headers: &mut HeaderMap)

Appends some entries into the header map of an HTTP response.