Module tsukuyomi::output

Components for constructing HTTP responses.

Re-exports

pub use tsukuyomi_macros::Responder;

pub use izanami::http::body::Body as ResponseBody;

pub use izanami::http::response::IntoResponse;

pub use izanami::http::response::Response;

Structs

Html

Json

JsonPretty

Oneshot

Rendered
A Responder that uses the specified preset.

ResponderFn

Traits

Preset
A trait representing the preset for deriving the implementation of Responder.

Respond
The asynchronous task that generates a reply to client.

Responder
A trait that abstracts the "reply" to the client.

Functions

html
Creates an HTML response using the specified data.

json
Creates a JSON responder from the specified data.

json_pretty
Creates a JSON response with pretty output from the specified data.

oneshot
Creates a Responder from a function that returns its result immediately.

render
Creates a Responder using the specified preset and data.

respond
A function to create a Responder using the specified TryFuture.