Struct tsukuyomi::server::transport::Builder

pub struct Builder { /* fields omitted */ }

A builder for constructing a Listener.

Methods

impl Builder

pub fn bind_tcp<A>(&mut self, addr: A) -> &mut Builder where
A: Into<SocketAddr>,

Configures to use a TCP listener that will bind to the given listener address.

pub fn bind_uds<P>(&mut self, path: P) -> &mut Builder where
P: Into<PathBuf>,

Configures to use a Unix domain listener that will bind to the given path.

NOTE: This method is enabled only on Unix platform.

pub fn finish(&mut self) -> Result<Listener, Error>

Finishes the current building session and constructs an instance of Listener with its configuration.