Trait tsukuyomi::Extractor

pub trait Extractor {
    type Output: Tuple;
    type Error: Into<Error>;
    type Extract: TryFuture<Ok = Self::Output, Error = Self::Error>;
    fn extract(&self) -> Self::Extract;
}

A trait abstracting the extraction of values from the incoming request.

Associated Types

type Output: Tuple

The type of output value extracted by Extract.

type Error: Into<Error>

The error type that may be returned from Extract.

type Extract: TryFuture<Ok = Self::Output, Error = Self::Error>

The type representing an asynchronous task to extract the value.

Loading content...
Required methods

fn extract(&self) -> Self::Extract

Creates an instance of Extract.

Note that the actual extraction process is started when the value of Extract is polled.

Loading content...