Trait tsukuyomi::Responder

pub trait Responder {
    type Upgrade: Upgrade;
    type Error: Into<Error>;
    type Respond: Respond<Upgrade = Self::Upgrade, Error = Self::Error>;
    fn respond(self) -> Self::Respond;
}

A trait that abstracts the "reply" to the client.

Derivation

The custom derive Responder is provided for reduce boilerplates around trait implementations.

The macro has a parameter #[response(preset = "..")], which specifies the path to a type that implements a trait Preset:

use serde::Serialize;

#[derive(Debug, Serialize, Responder)]
#[response(preset = "Json")]
struct Post {
    title: String,
    text: String,
}

You can specify the additional trait bounds to type parameters by using the parameter #[response(bound = "..")]:

#[derive(Debug, Responder)]
#[response(
    preset = "Json",
    bound = "T: Serialize",
    bound = "U: Serialize",
)]
struct CustomValue<T, U> {
    t: T,
    u: U,
}

Associated Types

type Upgrade: Upgrade

The type of asynchronous object to be ran after upgrading the protocol.

type Error: Into<Error>

The error type that will be thrown by this responder.

type Respond: Respond<Upgrade = Self::Upgrade, Error = Self::Error>

The asynchronous task converted from this responder.

Loading content...
Required methods

fn respond(self) -> Self::Respond

Converts itself into a Respond.

Loading content...