Trait tsukuyomi::upgrade::Upgrade

pub trait Upgrade {
    fn poll_upgrade(&mut self, io: &mut Upgraded) -> Poll<(), Error>;

    fn close(&mut self);
}

A trait that abstracts asynchronous tasks to be ran after upgrading the protocol.

Required methods

fn poll_upgrade(&mut self, io: &mut Upgraded) -> Poll<(), Error>

Polls the completion of this task with the provided I/O.

fn close(&mut self)

Notifies that the task is to be shutdown.

Loading content...