Trait tsukuyomi::util::TryFrom

pub trait TryFrom<T>: Sized {
    type Error: Into<Error>;
    fn try_from(value: T) -> Result<Self, Self::Error>;
}

A trait representing the conversion from the arbitrary types.

This trait is an emulation of the standard TryFrom.

Associated Types

type Error: Into<Error>

Loading content...
Required methods

fn try_from(value: T) -> Result<Self, Self::Error>

Loading content...